Optimize bare-centos-7 and rdo-centos7-liberty images
- Remove usage of EPEL (setuptools -> easy_install -> pip -> tox)
- Don't install things like khaleesi and ansible by default
- Remove duplicated install of packages
- Use centos-release-openstack-liberty package instead of
  setting up repository ourselves

It was necessary to "fork" base.sh and rpmfactory-base.sh as
they are part of software factory and any changes to them would
be overwritten by an update.
base.sh was forked mostly to get rid of setup using EPEL.
The only reason for rdo-rpmfactory-base.sh to exist is because
the default one imports base.sh which we want to replace by
rdo-base.sh.

Change-Id: I0f69f47dc033bf6a6af1ebedf71cf6b89d5607bf