Update patch set 4
Patch Set 4:

Just a note about pkg-export. This one was supposed to be run only in the gate pipeline to do a non scratch CBS build (so it make sense only in the gate) and one really important point is that job waits for others jobs started (in the zuul queue) related to that change to finish with success to do the non scratch build. I don't think Alan want to use that feature ATM for stable branches and prefer to do the CBS builds by hand.

Patch-set: 4