Update patch set 4
Patch Set 4:

@Tristan Thanks a lot !
The puppet-openstack voting was a good catch and was a typo. There is a "voting: false" for the puppet jobs at the top but the names didn't match, this is now fixed in latest patchset.

I have removed pkg-export from the check pipeline.

Don't know if it's good now, would like to have fbo confirm.

Patch-set: 4