Initial commit to activate the stable branch build on CBS
According to the card https://trello.com/c/URAtrhLU/86-automate-stable-packages-releases
this patch refresh rpmfactory pre-existing (but not used yet) jobs.

A follow up patch will be needed to set the UUID of the credential we will
use for CBS as well as adding the job a project we select as candidate for
the experimentation.

Change-Id: I0f18e06ab4645f24eb07c3512c9046600b356a00