Update patch set 2
Patch Set 2:

David, I don't know, I'm not aware of those jobs what the goal of them ? Could you explain. This patch is suppose to activate a check job that will build the stable branch package on CBS and basically test and install the package(s) on the slave node. There is also the second phase (not activated in that patch) in the gate pipeline where the package is build in the non scratch mode in order to land on the CBS repo.

Do you you think we can find synergy between both kind of job ?

Patch-set: 2
Label: Code-Review=0