Update patch set 3
Patch Set 3:

Are we talking about zuul/integrated.yaml and the artifact exchange between DLRN-rpmbuild and the child jobs ?

If that's the case then I'm not sure we can reuse because in the case described on the Trello card we want the build from CBS right (not from DLRN) ?

For now here we talk about just one job that will asks a build on CBS then perform a test on the built package(s) so there is no need for an exchange of artifacts. But if we want to run multiple jobs, and re-use the parent/child jobs technique introduced by David then we will need to store the Koji TaskID somewhere (in swift, under LOG_PATH maybe ?). But I think that's out of the scope of that patch now. We should validate the behavior with a single job and decide later if you need more job and at that time think about how to exchange the TaskID.

Patch-set: 3
Reviewer: Gerrit User 112 <112@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0