Update patch set 2
Patch Set 2:

This succeed https://review.rdoproject.org/r/#/c/1782/, just the listing of the packages from the temporary repository was broken. This patch fixes it and also install the built packages on the slave node. This should finalize the first part of the related Trello card.

Patch-set: 2
Label: Code-Review=0