Activate experimental CBS non-scratch build in the gate pipeline
This patch activates the job newton-distgit-cbs-final-build in the gate
pipeline. The job will run for only python-keystoneclient and dib-utils
for the newton-rdo branch.

This job will, if there is a NVR change between HEAD and HEAD^1, start
a non-scratch build on the given newton CBS target. If there is
no change the job just returns a success.

The patch I93f49cf6a75dc110f6c87cd173e5984af6ed113c needs to be approved
and merged before merging this one.

This is part of the work on the trello card
https://trello.com/c/URAtrhLU/86-automate-stable-packages-releases

Change-Id: I86c46dfd3db912a5b33d8ec49d1da18479689d80