Update patch set 2
Patch Set 2:

More similar to what's going on upstream. This is just a proof-of-concept to see if multinode is working as expected for R.R.O. After this merges I'll create a new test job on a repo that does something like setup a vxlan VPN between them and do a ping test.

The plan is to test quickstart using tripleo.sh's setup_multinode, then use work similar to slagle's -oooq option. It will likely have to be modified a bit to set up the correct RDO repositories for testing.

More details can be found in the latest comment of this Trello card: https://trello.com/c/lCvVr7QQ

Patch-set: 2