Update patch set 2
Patch Set 2: Code-Review-1

This isn't going to work because the DLRN-rpmbuild (and by extension, run_tests.sh [1] from DLRN) is designed to work against a distgit repo review.

If we want to do this, we're going to either adjust run_tests.sh to be able to support reviews on a project repo (not distgit) or create another script (i.e, run_project_tests.sh or whatever) that would map the project to the distgit repo accordingly.

We also want to zuul-clone the upstream project source and have dlrn build from that source, so I think we also need the --local flag on the dlrn command.

[1]: https://github.com/openstack-packages/DLRN/blob/master/scripts/run_tests.sh#L18

Patch-set: 2
Label: Code-Review=-1