Consider review.o.o projects when preparing workspace
Previously, prepare-workspace always assumed that projects where
hosted in review.rdoproject.org. Now that we can also act as a 3rd
party CI for review.openstack.org, we need to keep that in mind and
clone the repository from the OpenStack git if needed (that is, when
the pipeline starts with openstack-).

Change-Id: I543848a23ad5a3e5af26665ee81180ffa26a2630