Create RDO-specific prepare-workspace builder
When review.rdoproject.org is acting as a 3rd party CI for OpenStack
projects, we need a specific workspace preparation (use zuul-cloner
with the upstream OpenStack repos instead of review.rdo.o).

In https://review.rdoproject.org/r/2086, we did this by modifying
the sf_jjb_conf.yaml, which is a "provided by default" file. Here
we are creating a specific builder in rdo.yaml, and using it for
DLRN-rpmbuild.

Change-Id: I3068cabc60f792fea63b57dba7f92609b895a8be