Send dlrn-rpmbuild jobs to trystack, disable weirdo integrated jobs
rcip-dev is having performance issues right now causing jobs that are
run on it to time out and we are unable to troubleshoot it in a timely
manner.
trystack can't run weirdo jobs due to lack of nested virt so let's
disable the integrated jobs for the time being.

This should be addressed and reverted ASAP...

Change-Id: Id899cece3d4706d3fcb2cc4d2f0d42dcee6b0a8f