oooq: add dlrn package builder, fix tripleo-ci multinode setup
This commit adds the dlrn package builder to the tripleo-quickstart
jobs. It also installs a yumrepo, and does so on all of the subnodes.

The fix to tripleo-ci is that the repo is expected to be in a certain
place on each subnode. This commit ensures a clone matches that.

Change-Id: I08f21bb969000ee812d9a56edda8bdaf92aec919