oooq: add undercloud-post, manually run overcloud deploy, heat template paths
This commit splits out the overcloud deploy since we need to modify the
deploy script to disable the undercloud resource check loop. This also
adds the undercloud-post to create the final network-environment.yaml
heat template and adds a tripleo-ci checkout for use by the stack user
because the jenkins-owned checkout is unreadable due to permissions problems.

Change-Id: Ie67f67a21cc5557a5400305f0485fe5eb51a6d49