oooq: add overcloud-prep-image role and dependencies
This commit adds the overcloud-prep-image role and makes adjustments for
scripts to be able to deploy on nodepool. Specifically, it adds a
resource for adding instackenv.json lifted from the libvirt/overcloud
role and a modified prep-image script negating the ironic calls, since
these are not used on a nodepool deployment.

Change-Id: Ibd266f724b1da493291c9e258cf65852b8142792