oooq: modify heat templates, remove unnecessary prep roles
This commit modifies the heat templates passed to the job by adding the
deployed-server-environment template. It also removes the unnecessary
overcloud-prep roles and adjusts the arguments to the
overcloud-prep-images role to simply upload the overcloud image to
glance.

This commit also adds the COMPUTE_HOSTS and CONTROLLER_HOSTS environment
variables for the heat template to use.

Change-Id: I89391f387c622ce51043e4ef9ecfac4161f81c89
Depends-On: If59eac2f5df17a52a367c85546aa2946479df368@gerrithub