oooq: Update overcloud-deploy execution with full path, fix logging
The overcloud-deploy.sh command is issued through 'sudo -Eu'. The -E
flag means to copy over the environment variables, which includes
PWD/CWD. This means in order to execute the script we should either 'cd'
into the correct dir, or issue a full path.

This commit also removes an extraneous builder that produces a list of
installed RPMs.

This commit also simplifies the collect-logs builder by using the
$WORKSPACE variable instead of issuing ansible to perform a lookup()

Change-Id: Iea556c2208c430a9a84553a498043a30ef5198aa