Update list of pre-installed packages for weirdo jobs
The original implementation was flawed and packages that were
installed from RDO repositories were mistakenly pre-installed
from CentOS base repositories and this was causing conflicts.

We now properly filter the packages we install such that we do not
install any conflicting packages and we also remove any packages
that could potentially be installed from a dependency.

As of this current commit, this results in pre-installing 386
packages that are required throughout the different jobs, saving
a couple minutes in the process.

Change-Id: If8b7ae46083168f907444369425e696983ce35de