First proposal of resources refactoring
This patch to bootstrap a new file resources/rdo.yaml where
we could describe RDO's groups, acls, repositories and
a meta RDO project.

Please have a look to the config-check logs in order
to get details about actions that will be taken at
config-update level when that patch will be approved.

sf-resources: allow-delete

Change-Id: I7051cc45a41c79eccb01165ddf2e828e9d3e7452