Update patch set 2
Patch Set 2:

After the upgrade a tool has run to extract all repo, groups, acls from Gerrit and a file missing_resources_1489486510.yaml have been created.

But as you can see this is a big raw file. Please note that included group have been discarded from the extract as not supported by the new resources engine.

This patch propose the bootstrapping of a file rdo.yaml with the migration of barbican* resources into it. For both projects barbican and barbicanclient only two acls are defined:
* rdo-upstream-mirror
* rdo-distgit
and only two groups are referenced into them rdo-provenpackagers and rdo-superusers. I guess most of packaging project can just relies on those acls and groups and this will allow us to have a more clear view of the acls state.

But this is just a proposal feel free to comment or change. Please check also the config-check log as the consistencies of the change is tested there and also clear log of events (that will happen when the merge occur) are given.

And finally I prefer if we can iterate by migrating a limited set of resources from the missing_resources.yaml to the new rdo.yaml file to avoid side effect. Migrating so much resources is quite exceptional and is not suppose to occur again.

Patch-set: 2