Update patch set 2
Patch Set 2:

> I'm afraid that having a simple set of acls will not work for us,
 > since we want to give each contributor control over their own
 > packages. We may keep the rdo-upstream-mirror/rdo-distgit acls for
 > generic projects without a defined maintainer, though.
 > 
 > Thinking about the structure, is it possible to split the
 > definitions into different files, like:
 > 
 > - acls.yaml
 > - groups.yaml
 > - repos.yaml
 > 
 > That would allow us to document a sample block to use for each
 > file, simplifying the creation of new projects.

What about having a file for each set of functinally related projects?, for example:

- rdo-puppet.yaml
- rdo-infra.yaml
- rdo-glance.yaml
- rdo-neutron.yaml
- etc...

It'd be easy to create a template for the entire file that could be used when adding new projects, and it'd require to modify/create a single file, not modifying three of them. We could end up with a good number of files, i'm not sure if that could be a problem.

Patch-set: 2
Label: Code-Review=0