Update patch set 6
Patch Set 6: -Code-Review Workflow-1

There is already a collectd repo at github.com/centos-opstools, so as soon as this is merged it could be wiped with an empty repo.

We need to figure out how to exclude the repo from synchronization until the initial import is done.

Patch-set: 6
Label: Workflow=-1