Update patch set 2
Patch Set 2: Code-Review+2 Workflow+1

The rdoinfo project creation only adds the resources. This, in turn, creates the gerrit project.

If the gerrit project exists but is not mapped in gerrit replication, gerritbot or zuul, this will cause the config repo sanity check job to fail after the patch has merged because there is now an "orphan" gerrit project.

I feel like everything should be in the same review but it's hard to automate because there might be "logical" decisions, especially around Zuul layout.
This is okay for now, let's revisit this when jpena is back from pto.

Patch-set: 2
Reviewer: Gerrit User 16 <16@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+2
Label: Workflow=+1