Use ZUUL_REFNAME to clone from in prepare-workspace for post jobs
We are finding issues in some rdoinfo post jobs, when we have a merge
commit. This is because we are trying to clone ZUUL_NEWREV, when the
merge commit may not be there in the Zuul local cache. For those
cases, we should use ZUUL_REFNAME as the reference to clone.

Change-Id: I6799a94d6f5f1d54b96cb31a0e715f947d69618f