Update patch set 2
Patch Set 2:

> Not sure to understand, zuul_ref already equals zuul_refname in
 > post pipeline: https://review.rdoproject.org/jenkins/job/rdoinfo-create-project/38/parameters/
 > 

Yes, but if there is a ZUUL_NEWREV environment variable, it tries to use it as reference unless we specify it in the command line, or at least that's what I saw in my tests.

 > Moreover, looking at a past job, e.g. https://review.rdoproject.org/jenkins/job/rdoinfo-create-project/30/parameters/,
 > zuul_newrev is referencing a merge commit and the job succeeded.

Yes, that's strange. Maybe there is something wrong with the local copy? Looking at https://review.rdoproject.org/jenkins/job/rdoinfo-upper-constraints-update-master/lastFailedBuild/console (a periodic job), it didn't use the latest commit in the repo, but a previous one. Maybe there's an issue with the Zuul local cache?

Patch-set: 2
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0