Automate new-version updates using post pipelines
We can trigger new-version automatically in two cases:

- A new release is tagged upstream using openstack/releases
project and is not pinned in rdo.yml.
- Version is updated in rdo.yml for a package pinned in rdoinfo.

This patch adds a new job rdo-send-stable-release that creates
reviews for new versions based on these two scenarios and adds the
job to post pipeline of both projects.

Note that i'm running rdo_release_revew with --dry-run to test it
doesn't break anything.

Change-Id: I680e751975fe3d767ceb6375aa7a1fe3b32b7ef3