Add option to skip weirdo job if a flag file exists
When automating buildsys tagging the weirdo jobs to be
executed will depend on the repos changed by commits in
rdoinfo.

Zuul has no way to define the jobs to be executed depending
on the string changes in the review (only in the changed files)
so the plan is to run a builder that will configure temporary
repos before running weirdo and when a job is not required it
will create a flag file.

This patch uses that flag file to skip the test when it's not
needed.

Change-Id: I60295577420ab4ab3a8750e564e62978cfd62616