Update patch set 4
Patch Set 4:

> Is it worth it to bump to the flavor with 8GB of RAM ?

It's only happening in one project so far (Cinder), and in that case the delta between using tmpfs and not using it is 3 minutes out of 16. I think we can keep it within 4 GB for now, and bump only if needed.

Patch-set: 4
Reviewer: Gerrit User 113 <113@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0