Add credential-binding with weirdo token for ci-config project
Some jobs managed in ci-config project (running in ci.centos.org)
need to have an auth-token so that we can trigger them remotely
in a secure way.

This patch adds a credentials-binding for a file which contains
the desired token and copies it to location jenkins/jobs/weirdo_token
to be included in ci-config project using:

auth-token: !include: 'weirdo_token'

Additionally this patch removes from printing to stdout any message
containing the strings auth-token and authtoken to avoid leaking the
password in the job console logs.

Change-Id: I6a164bf2239f272eac0e41ff96df79473eb580bc