Update patch set 1
Patch Set 1:

I have some doubts here. If we run all jobs with SELinux=permissive, we will never be able to find/fix issues that are SELinux-related.

With openstack-selinux being built in RDO Trunk (and thus present in the promotion pipeline), shouldn't we do it the other way around and always run in enforcing mode? Otherwise, RDO users will be the ones finding SELinux-related problems in test days or post-GA.

Patch-set: 1
Label: Workflow=0