Update patch set 1
Patch Set 1:

> I have some doubts here. If we run all jobs with SELinux=permissive,
 > we will never be able to find/fix issues that are SELinux-related.
 > 
 > With openstack-selinux being built in RDO Trunk (and thus present
 > in the promotion pipeline), shouldn't we do it the other way around
 > and always run in enforcing mode? Otherwise, RDO users will be the
 > ones finding SELinux-related problems in test days or post-GA.

is selinux enforcing mode in release criteria for RDO?, my understanding is that it isn't. If it is, we should fix it in promotion pipelines. There may be intermediate solutions as enforcing only in packstack jobs but leaving permissive for p-o-i (which forces it to permissive) and oooq.

Patch-set: 1
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0