Update patch set 2
Patch Set 2: Code-Review+2

OK for now but as per discussed, we'd need both packstack and p-o-i to support running in:
- Permissive while catching AVCs (current p-o-i behavior)
- Enforcing with hard failures

Testing selinux enforcing upstream is valuable as it allows us to catch problems much earlier in the production chain instead of unnecessarily burden downstream on *finding* those.
It doesn't mean that it needs to be voting (i.e, part of the promotion pipeline) but with DLRN API, enforcing jobs reporting their status would be nice feedback.

Patch-set: 2
Label: Code-Review=+2