Update patch set 2
Patch Set 2:

BTW, we're running the Packstack upstream gate with SELinux=enforcing. Would it be ok if we add some login in the Packstack check scripts to make this configurable, but still have enforcing by default?

This would give us the same behavior for Packstack everywhere (enforcing by default), and the same for p-o-i (permissive by default).

Patch-set: 2