Update patch set 2
Patch Set 2:

> Fabien, can you comment on this? I think that, for an existing
 > project, we could simply add the branch using the Gerrit UI, but
 > I'm not sure if that's the best method.

Note that one of the advantadges of this is to allow project maintainers to propose branches by themselves. I want to test it with this change so that we can use it to branch pike at rc1 timeline.

Patch-set: 2
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0