Add rdo-cloud-tripleo nodepool provider, move upstream-centos-7 image
This leverages the 'openstackzuul' tenant in RDO Cloud for the purpose
of running jobs based on tripleo.
Jobs from that tenant can and will create VMs in the 'openstack-nodepool'
tenant for OVB purposes. The VMs inside the 'openstack-nodepool' tenant
are ironically *not* managed by nodepool, the 'openstack-nodepool' tenant
is named thusly due to expectations from tripleo/tripleoci script expectations.

We're also moving the upstream-centos-7 image to that new provider, it's
where it is meant to be used.

Change-Id: Icd74f6ee64624f910894caf240b9acbc1605bd43