add periodic jobs for upstream tripleo
We'll start adding two jobs:
- a know to work job that tests ovb ha
- a new job to test an promote containers images.
we're putting them in a separate pipeline called "manual"
the periodic job will sometimes need manual triggering, for example when
we want to force a promotion.
This job will be initially launched from an upstream review in
openstack/tripleo-ci, by a "check manual" comment.
When we are sure they're working, we'll create the periodic pipeline and
add (not move) the job to it.

Change-Id: I8439aa03da20a177ac33d0be8557b3b25f4df96f