Add bot to update buildsys tags automatically
This change adds some jobs to periodic pipeline to propose
changes on rdoinfo to update buildsys-tags from candidate to
testing and from testing to release, for newton and ocata.

Note that this is not updating dependencies nor changes in
pike testing as we still have some work to do to move BR only
out of candidate tags.

Change-Id: I63ec64ba2199ef0b2c8de6065d70416f7eea08fe