Update patch set 2
Patch Set 2: Code-Review+1

This all looks good to me..  The job name should not be based on the cadence, but the pipeline name should be.

I am not sure if we'll be able to build/test w/ containers in 4hr so I could see that slipping to 5 or 6 hours.  Still much better than every 24hr.

Patch-set: 2
Reviewer: Gerrit User 138 <138@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+1