Create and test new DIB CentOS 7 image
We're not going to use the software factory elements for the time
being because they, amongst other things, install packages and
repositories that could conflict with jobs we will be running.
We want an image that is as vanilla as possible.

This new image is essentially the same as upstream-centos-7 minus
- Devstack cache (opentack-repos/cache-devstack)
- Preinstalled packages (infra-package-needs)

Change-Id: I063c54707669bce15b54688823a249cf6062313a