Move all jobs from bare-centos-7 to rdo-centos-7
The new rdo-centos-7 DIB has shown to be successful in replacing
the bare-centos-7 image in different spot checks.
We need to move away from the bare-centos-7 image as it is based on
snapshots and these will no longer be available in the upcoming
upgrade of Software Factory.

Modepool capacity is shuffled according to this new change.
We are getting rid of the version with 4vCPUs and 4GB RAM for the
sake of simplicity for the time being.
We can re-introduce it later if necessary -- for now it grants the
ability to spawn up to 33 concurrent nodes with the current quotas.

Note that we're still binding weirdo jobs to rcip-dev until boot
from volume is available in nodepool with Software Factory 2.6.
RDO cloud ephemeral storage I/O performance is problematic when
running OpenStack integration jobs.

Change-Id: I12ede44289efe1af724ccbabdb78439f9b495ad9