Rename ansible-role-rdobase to ansible-role-rdo-base
Not sure why it was named rdobase in the first place and it's been
bothering me too much.
This should be perceived as a new project in gerrit, which is fine,
we'll delete the previous one.

Actually renaming a project in gerrit is a lot of trouble for little
value for this small project.
We're also homogenizing the ACL for rdo-infra projects.

sf-resources: allow-delete
Change-Id: Id4cb8315f6da25a187ab50d97a35a8341a9717aa