Configure prepare-workspace to use git clone for periodic jobs
periodic jobs aren't properly namespaced against the project they
are configured on.
Even if a periodic job is triggered against an upstream project,
it will be namespaced under review.rdoproject.org which is a problem
because, for example, review.rdoproject.org does not carry
"openstack-infra/tripleo-ci", so zuul-cloner won't find the repository.

There's no real reason to be using zuul cloner when running periodic
jobs anyway, so let's fallback to regular git clones until we find
a better solution.

Change-Id: Ibf70cde4f61be79c640d833ff3e7e2f8ba257b2a