Add third-party CI on tripleo-common to test container builds
This adds a job that will run DLRN-rpmbuild on new tripleo-common
reviews and then run a container build with the new package to ensure
that new images or overrides do not break container builds.

This doesn't test the containers themselves but at least we'll know
if they can build, we'll eventually add further testing coverage.

Change-Id: Ib80dfb10573255321f20132d76e42c53d58c6c0a