tripleo=upstream: add two experimental pipelines
WAT ? we don't need two more pipelines, WTH !
Maybe not, but let me try to explain.
I need to test two different stages of the same job: the parent that
builds images, and the children that test them.
Now that I know that the parent works, i don't want to relaunch it every
time just to test the child. I need two different triggers for them.
One will spin everything, the other just the children.
Why the move from experimental to ci-experimental ?
experimental is very common in upstream, and I think it's used for both
experimenting new jobs or set of jobs or pipelines (so from a infra POV)
and users that need to test experimental features in some non voting
jobs.
It already happened that someone comments "check experimental" in a
tripleo-ci change to test the upstream experimental jobs, and the
rdocloud side gets launched, and since it's promoting, it changes repo
links.
I just want to split the two sides (infra experimentation and user
eperimentation) in two differently named pipelines.
let me know if you can find a better soluition for me.

Change-Id: Idbb891c693c11aa8924ac091baf1a193c2923e37