Drop mirror creation from upstream-centos image
The mirror information for upstream is being put into the upstream
centos image which is problemattic because if your job uses this
information it will fail.  We can build the image but we should not
include the mirror information.  See failures on the upstream change
https://review.openstack.org/#/c/492289/ where the rax mirror
information was being used by the RDO 3rd party CI.

Change-Id: I6c6fef0a745188da97577c9133c1baa69c38137e