Use in progress tripleo-ci review for new tripleo periodic job
We currently have the new periodic job working in so far as it
is capable of building images/containers then running child jobs
that test them. However, the tripleo-ci review needs quite a bit
of cleanup to be mergable.

At the same time, the containers test is failing because we need a
promotion to fix gnocchi. As such, it would be good to be actually
running the periodic job to validate that fix when it lands while
we work to clean up and merge the tripleo-ci review.

Change-Id: I968e7b071f06f00568c8e74e702cbb3388754d50