Clone to the same location for all jobs
We have a need to clone using zuul-cloner in some cases while just
using git in others. However, the current implementation clones to
a different location depending on which method is used. This patch
standardizes the git method to clone to the same location as
zuul-cloner.

This is important to be able to use the same builders (which rely
on some checkout in the workspace) to be used for all the different
job types.

Change-Id: I0577e63f14d1f13aaf425be5abefe694b7fdb0eb