Update patch set 2
Patch Set 2: Code-Review-1

This package has not been added to rdoinfo and there is no package review bugzilla. The process to add packages to RDO is in:

https://www.rdoproject.org/documentation/add-packages/

Said this, this project is not under openstack umbrella but in kubernetes (https://github.com/kubernetes-incubator/kubespray) so i'm not sure what's the best way to manage this. Could you provide some more background about the relationship of this with OpenStack?, will be this required by any OpenStack project?

Patch-set: 2
Reviewer: Gerrit User 128 <128@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=-1