Run tripleo-ci/toci_gate_test.sh from our git checkout
Currently we are running the main CI script from /opt/stack/new/
This works fine for the zuul cloner case, and would likely work
fine once we no longer need an in progress tripleo-ci review.

However, this prevents us from using the git checkout with our
in progress review, and it is actually a bit unclear to me how the
/opt/stack/new dir is being populated.

It is better to just run everything from our git checkout where we
know exactly what we have.

Change-Id: Icb02cf13f8c9a2d62fb11630277b5434dd16011d