Add OVB jobs and 24hr pipeline
This is review enables OVB periodic jobs and assigns
jobs to be triggered by either the 8hr (master, pike)
or the 24hr (previous releases) pipelines.

The manual pipeline is meant for use in manual promotions
only. An experimental pipeline is added to check jobs
under test.

This review also increases max-server nodes for
tripleo-upstream RDO provider to 50.

Change-Id: Ide01ef8ff414a8c6b90ccc1769f77d8dbf69a2ed