Use unique comment to trigger rdoproject experimental jobs
There is already a pipeline in openstack-infra that is triggered
by "check experimental". We currently use the experimental pipeline
in rdoproject zuul to iterate on the periodic container build and
validate job. However, if some one tries to check experimental on
some other tripleo-ci patch (because they want to run
openstack-infra experimental jobs), it will cancel our run and we
will have to restart it.

Instead, we can use a unique name for our experimental pipeline.

Change-Id: Ief8b5cccf08cd741c2f270df5e665ee9b4c6d8c7