Set options for rotate and timeout in rdo-base element
We can reduce the DNS load on a single server by setting the rotate
option in resolv.conf [1]. The current configuration for the RDO CI
tenant includes 4 servers, so using it can spread the load and
minimize the chance of being rate-limited.

Also, setting a short timeout should allow us to skip servers with
temporary issues.

[1] - man7.org/linux/man-pages/man5/resolv.conf.5.html

Change-Id: Iaf1a1e3088e87f171bcef3fa758fa0b6fd3dde0d