Add buildsys-tags jobs for pike release
This change implements required changes in job configs for
new pike release:

- Adds tripleo jobs for pike CloudSIG testing and release repos
- Adds bot rdoinfo-buildsys-tags-update-pike-{testing,release} to
  automatically propose buildsys-tags updates in rdoinfo with pending
  builds

This commit must not be merged until all CBS builds for pike are created
and in pike-candidate.

The jobs for pike-release are created but are unused (commented out) in
the Zuul layout, they should be enabled once Pike is actually released.

Change-Id: Ib0055943dde37c4eeb0731e56f2f72097dcaccc1