Remove qcow image upload from container build job
We need to stop building the qcow images in the multinode jobs,
because it breaks tripleo upstream[1][2]. A follow-up patch
will make one of the OVB jobs upload its images. This will
also reduce the total job time, since it will reduce the time
of the container/build test job.

[1] https://review.openstack.org/498772
[2] https://launchpad.net/bugs/1713167

Change-Id: I24608903e197f9a644d1ec8b0a8816172a8a3959