WIP - opstools-ansible-distgit project
To create the opstools-ansible rpm package for CentOS.
Disable the current rpm check from the opstools-ansible
repo, and prepare the new for opstools-ansisble-distgit. This
is no longer need it.

Modified the jobs base on the spec files not to base on
the name of the repository:
- Currently the jobs use the $reponame.spec to check and
creates the packages

- The patch do the same and if the $reponame.spec does not exit
try to find the first spec file on the repo.

Change-Id: I1a87267138ad8767c0b186d8f4f30cbd1a7e776a