Update patch set 1
Patch Set 1:

Sagi, I'm not saying I'm blocking the upload of the console log. I'm saying I'm blocking it through the use of that particular publisher.

If this is required in the short term, you will need to implement a different way of uploading the console log. For example on ci.centos.org we do it this way: https://github.com/rdo-infra/weirdo/blob/master/playbooks/logs-ci-centos.yml#L30

Patch-set: 1