Add new "virtual" provider ssd-rdo-cloud, assign a test job
We are going to test a different aggregate in the RDO Cloud, using
local SSD disks instead of Ceph, to see if this helps alleviate our
performance issues.

We do so by creating a new provider (ssd-rdo-cloud), based on the
same cloud credentials as rdo-cloud, and a label (ssd-rdo-centos-7)
based on the rdo-centos-7 image. To test, we will assign the
rdoinfo-DLRN-check job to this label/cloud.

Change-Id: I62bed324fd8611c9185eaaa54ec5bcb4a6cd3b03