Add new rdo-cloud-ssd nodepool provider
We are going to test a different aggregate in the RDO Cloud, using
local SSD disks instead of Ceph, to see if this helps alleviate our
performance issues.

We do so by creating a new image label (rdo-centos-7-ssd) for the
rdo-nodepool tenant, which uses the appropriate flavor. To test, we
will assign the rdoinfo-DLRN-check job to this image.

Change-Id: I62bed324fd8611c9185eaaa54ec5bcb4a6cd3b03