Move weirdo jobs in review.r.o gates to SSDs
Currently weirdo jobs are being executed in instances
with ceph backed cinder disks but this is causing performance
issues and jobs are extremly un-reliable.

This patch moves weirdo jobs to servers with local SSDs
and re-balance the rate between max number of instances
with and without SSDs to manage all these jobs.

Change-Id: Ibb5e11f2a386c46dab4b1e86d796457090d83777