Deploy puppet modules from RPMs
Currently, in p-o-i jobs running in review.rdoproject.org
we are deploying puppet modules using Puppetfile in the
upstream repo, so jobs are not valid to test bumps in
puppet modules.

This patche enables puppet modules deployment from RPM so that
we can use it to validate bumps on them.

Change-Id: Iffebfd541ad6e01e6d3fc25355e35dfa15db694e