Update patch set 1
Patch Set 1:

> (1 comment)

what we're trying to do is communicate to reviewers how to treat this job.

afaict, it's failing all the time right now (based on the handful of patches I've spot checked and reviewed).

Does that mean...don't merge any code? Is the job expected to be passing or failing? Who knows? Who is maintaining the job? It's not up to reviewers to fix it. Who should be fixing it?

A job with a -nv suffix sends a clear signal to reviewers how to treat the results. My concern with this job initially is that reviewers are already ignoring the results.

When I checked with Wes, even he was unsure. Sure, there was an email from 2 months ago...is all that still valid?

Patch-set: 1
Reviewer: Gerrit User 96 <96@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Workflow=0