[WIP] Create experimental job to test DLRN builds in Fedora
In some cases, we are finding it useful to build our packages for
Fedora Rawhide, for example when we are dealing with Python 3 or
with new dependencies being added. This patch adds a new job to
the experimental pipeline, where we can trigger a Fedora-based build.

The long-term plan would be to have it at least as non-voting, so
we can check when a spec is going to break in Fedora.

Note this requires https://softwarefactory-project.io/r/9926 to be
merged, hence the WIP.

Change-Id: Id4e734ecbde3ac11f0789c8cc0e5d595a175a35a