Update patch set 3
Patch Set 3: Code-Review+2

> @Matthias: I see some previous new packages had an associated bz in
 > RDO for package-review, but in this case they are related to
 > ovirt-distribution. I understand that, not being RDO packages, they
 > may require a different process, can you clarify?

Javier, I have no idea what this does or why it's needed.

That being said, opstools is following RDO package reviews. rubygem-faraday is in fedora already, there is no reason to deny this in opstools. If there's a usecase and a maintainer, we're fine to go with it.

I've created the repo on github.com/centos-opstools

Patch-set: 3
Reviewer: Gerrit User 80 <80@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=+2