Don't keep job artifacts forever, rotate 90 days
We have 19k directories worth of DLRN-rpmbuild jobs (amongst other
things) dating back to April 2016. This is costing us a lot of disk
space because not only does this take room, but we're also doing a
back up of those directories so it's twice the usage.

We don't realistically need to keep more than 90 days of *jenkins*
job history, let's rotate after that.

Change-Id: I500625a77cc78c17aba0719f0048433e78b1983a