Update patch set 2
Patch Set 2:

I've checked what I wanted to -- I wanted to make sure that setting the pipeline to high would not impact the RDO pipelines: http://eavesdrop.openstack.org/irclogs/%23zuul/%23zuul.2017-10-24.log.html#t2017-10-24T19:14:09

I'll defer from voting on this review, the configuration is good but the fact that you set the periodic pipeline to 'high' must be a TripleO CI one.

Pipeline priorities work this way: there's three precedence 'low', 'normal', 'high'.

If there is a pipeline with 'high' precedence, it means no jobs with a 'normal' precedence will be given a node so long as all jobs in the pipeline with a 'high' precedence do not have one.
Same thing if there is a pipeline with a 'low' priority, it means it won't get a node until *all* jobs in the high and normal precedence have been given a node.

Pipelines with the same level of precedence are dequeued on a first in/first out basis.

Patch-set: 2
Reviewer: Gerrit User 16 <16@270e2033-b340-4cff-9539-693957ebf0e7>
Label: Code-Review=0