Update patch set 1
Patch Set 1:

Let's discuss the numbers. It turns out there are 25 nodes available, instead of 15 as we originally thought (there was a typo in the documentation).

We're close to the limit in terms of CPU and disk with this configuration. Are we using this aggregate for anything else? We had some capacity being used by DLRN-rpmbuild jobs.

Patch-set: 1